feat(sledgehammer): CentOS 8 based Sledgehammer image.
This is a first-pass implementation of a CentOS 8 based Sledgehammer.
As of now, it boots and runs the discover-base workflow on amd64 boxes
in UEFI and legacy BIOS mode.

It still needs lots of testing and potential tweaking on all our
supported workflows before we can call it ready, but this is out there
for now for testing purposes.