wip(content,task-library) Ubuntu 20.04 Automated Install (incomplete)
This is the first pass at adding support for Ubuntu 20.04 automated
installs. The new Casper/Curtin/Subiquity based installer is very
different from the old preseed model from debian-installer as it
relies entirely on cloud-init. Memory requirements also massively
increased, a minimum of 3GB of RAM is needed on the target system.
The installation process is also very slow if there is no (or
partial) internet access, taking well over 5min to get up and
running as it has hard coded checks against Canonical's snap and
ntp services, among other things. Proxy settings are not checked
until after these hard coded checks are completed.

More information can be found at:
  https://ubuntu.com/server/docs/install/autoinstall

This is incomplete. The process is broken in the 20.04 install media
and the required testing tools listed on the above page do not exist
yet. The installer will boot, launch, but not do anything. This is a
known issue with the installer included in the install media. The
20.04.1 release should include the updated installer that should
fix this, and the development/debugging snaps should also be
available at that time. But this should allow for just replacing
the ISO and checksum in the bootenv to get working once the new
install media with updated installer code is available.